requirements.txt
torch
torchvision
pillow
pandas
matplotlib